fixes EOF issues and adds analytics to constants
