+++ src/js/util/analytics/actions.ts
@@ -1,0 +1,17 @@
+export const ACTIONS = {
+  SHARE: 'share',
+  LINE_SHARER: 'line-sharer',
+  GENERIC_ERROR: 'generic-error',
+  GURMUKHI_KEYBOARD: 'gurmukhi-keyboard',
+  COPY_SHORT_URL: 'copy-short-url',
+  SEARCH_QUERY: 'search-query',
+  SEARCH_SOURCE: 'search-source',
+  SEARCH_TYPE: 'search-type',
+  NO_RESULTS_FOUND: 'no-results-found',
+  ANG_NOT_FOUND: 'ang-not-found',
+  SEHAJ_PAATH_LINK: 'sehaj-paath-link',
+  RELATED_SHABAD: 'related-shabad',
+  JAPJI_SAHIB_LINK: 'japji-sahib-link',
+  REHRAAS_SAHIB_LINK: 'rehraas-sahib-link',
+  SOHILAA_SAHIB_LINK: 'sohilaa-sahib-link'
+};
