+++ src/js/util/analytics/page-view.ts
@@ -1,0 +1,8 @@
+/* globals ga */
+export const pageView = page =>
+  requestAnimationFrame(() =>
+    ga('send', {
+      hitType: 'pageview',
+      page,
+    })
+  );
