+++ src/js/util/array/index.ts
