+++ src/js/util/dom/index.ts
