+++ src/js/util/gurbani/index.ts
